+++ CHANGES.md
@@ -4,6 +4,8 @@ The release versions are PyPi releases.
 ## Version 3.5 (as yet unreleased)
 
 #### New Features
+  * added some support for extended filesystem attributes under Linux 
+  ([#423](../../issues/423)) 
   * added support for null device ([#418](../../issues/418))
   
 #### Infrastructure